implementing structure for integration tests
