+++ src/test/resources/hibernate.cfg.xml
@@ -1,0 +1,14 @@
+<?xml version = "1.0" encoding = "utf-8"?>
+<!DOCTYPE hibernate-configuration SYSTEM
+        "http://www.hibernate.org/dtd/hibernate-configuration-3.0.dtd">
+<hibernate-configuration>
+    <session-factory>
+        <property name = "hibernate.dialect">org.hibernate.dialect.PostgresPlusDialect</property>
+        <property name = "hibernate.connection.driver_class">org.postgresql.Driver</property>
+        <property name = "hibernate.connection.url">jdbc:postgresql://localhost:5432/TaskManager</property>
+        <property name = "hibernate.connection.username">root</property>
+        <property name = "hibernate.connection.password">root</property>
+        <mapping resource = "task.hbm.xml"/>
+        <mapping resource = "user.hbm.xml"/>
+    </session-factory>
+</hibernate-configuration>
